+++ app/models/logs.py
@@ -17,15 +17,16 @@ class Log(db.Model):
     timestamp = db.Column(db.DateTime, default=func.now())
 
     def save(self, **kw):
-        try:
-            db.session.add(self)
-            db.session.commit()
-            if self.id is not None:
-                return True
-            else:
-                return False
-        except DatabaseError:
-            pass
+        # try:
+        #     db.session.add(self)
+        #     db.session.commit()
+        #     if self.id is not None:
+        #         return True
+        #     else:
+        #         return False
+        # except DatabaseError:
+        #     pass
+        pass
 
     @staticmethod
     def load_by_level(level):
